feat: apply feedback_not use sharedPreferences directly in ViewModel
