+++ app/src/main/java/campus/tech/kakao/map/MainViewModel.kt
@@ -12,7 +12,8 @@ class MainViewModel(application: Application) : AndroidViewModel(application) {
 
     private val apiService = KakaoAPIRetrofitClient.retrofitService
     private val repository = PlaceRepository(apiService)
-    private val sharedPreferences = application.getSharedPreferences("search_prefs", Context.MODE_PRIVATE)
+    //private val sharedPreferences = application.getSharedPreferences("search_prefs", Context.MODE_PRIVATE)
+    private val preferencesRepository = PreferencesRepository(application.applicationContext)
 
     private val _searchResults = MutableLiveData<List<Document>>()
     val searchResults: LiveData<List<Document>> get() = _searchResults
@@ -47,24 +48,12 @@ class MainViewModel(application: Application) : AndroidViewModel(application) {
     }
 
     private fun loadSavedSearches() {
-        val searches = mutableListOf<String>()
-        val size = sharedPreferences.getInt("search_size", 0)
-        for (i in 0 until size) {
-            val search = sharedPreferences.getString("search_$i", null)
-            if (search != null) {
-                searches.add(search)
-            }
-        }
+        val searches = preferencesRepository.getSavedSearches()
         _savedSearches.postValue(searches)
     }
 
     private fun saveSearchesToPreferences(searches: List<String>) {
-        val editor = sharedPreferences.edit()
-        editor.putInt("search_size", searches.size)
-        searches.forEachIndexed { index, search ->
-            editor.putString("search_$index", search)
-        }
-        editor.apply()
+        preferencesRepository.saveSearches(searches)
     }
 
     fun searchSavedPlace(savedQuery: String) {
